structured_info/requirements.txt

openai
pandas
tiktoken
google-cloud-storage
requests==2.31.0
MechanicalSoup==1.3.0
beautifulsoup4==4.12.2
Flask==2.3.2
functions-framework==3.4.0
Pillow==10.0.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.1
img2pdf==0.4.4
python-magic==0.4.27
python-dotenv==1.0.0
pdf2image==1.16.3
supabase
openai>=0.10.2
azure-ai-formrecognizer>=3.1.0
pandas>=1.3.3
google-cloud-storage>=2.5.0
PyPDF2>=2.10.3
pandas>=1.5.2
google-auth>=2.6.0
functions-framework>=3.0.0
chardet==4.0.0  # or use whatever version you prefer